+++ core/src/main/scala/ai/h2o/sparkling/backend/api/ServletBase.scala
@@ -16,10 +16,22 @@
  */
 package ai.h2o.sparkling.backend.api
 
-import javax.servlet.http.HttpServlet
+import ai.h2o.sparkling.utils.ScalaUtils.withResource
+import com.google.gson.Gson
+import javax.servlet.http.{HttpServlet, HttpServletResponse}
 
 private[api] trait ServletBase extends HttpServlet {
   protected def toScalaRegex(pathSpec: String): String = {
     pathSpec.replaceAll("\\*", "\\.*")
+  }
+
+  protected def sendResult(obj: Any, resp: HttpServletResponse): Unit = {
+    val json = new Gson().toJson(obj)
+    withResource(resp.getWriter) { writer =>
+      resp.setContentType("application/json")
+      resp.setCharacterEncoding("UTF-8")
+      writer.print(json)
+    }
+    resp.setStatus(HttpServletResponse.SC_OK)
   }
 }
